+++ common/src/main/java/traben/entity_texture_features/config/screens/ETFConfigScreenBlinkSettings.java
@@ -13,6 +13,7 @@ import java.util.List;
 import java.util.Objects;
 
 //inspired by puzzles custom gui code
+@SuppressWarnings("AnonymousInnerClassMayBeStatic")
 public class ETFConfigScreenBlinkSettings extends ETFConfigScreen {
     protected ETFConfigScreenBlinkSettings(Screen parent) {
         super(ETFVersionDifferenceHandler.getTextFromTranslation("config." + ETFClientCommon.MOD_ID + ".blinking_mob_settings_sub.title"), parent);
@@ -25,9 +26,7 @@ public class ETFConfigScreenBlinkSettings extends ETFConfigScreen {
         super.init();
         this.addDrawableChild(getETFButton((int) (this.width * 0.55), (int) (this.height * 0.9), (int) (this.width * 0.2), 20,
                 ScreenTexts.BACK,
-                (button) -> {
-                    Objects.requireNonNull(client).setScreen(parent);
-                }));
+                (button) -> Objects.requireNonNull(client).setScreen(parent)));
         this.addDrawableChild(getETFButton((int) (this.width * 0.25), (int) (this.height * 0.9), (int) (this.width * 0.2), 20,
                 ETFVersionDifferenceHandler.getTextFromTranslation("config." + ETFClientCommon.MOD_ID + ".reset_defaults"),
                 (button) -> {