+++ common/src/main/java/traben/entity_texture_features/config/screens/ETFConfigScreenEmissiveSettings.java
@@ -22,9 +22,7 @@ public class ETFConfigScreenEmissiveSettings extends ETFConfigScreen {
         super.init();
         this.addDrawableChild(getETFButton((int) (this.width * 0.55), (int) (this.height * 0.9), (int) (this.width * 0.2), 20,
                 ScreenTexts.BACK,
-                (button) -> {
-                    Objects.requireNonNull(client).setScreen(parent);
-                }));
+                (button) -> Objects.requireNonNull(client).setScreen(parent)));
         this.addDrawableChild(getETFButton((int) (this.width * 0.25), (int) (this.height * 0.9), (int) (this.width * 0.2), 20,
                 ETFVersionDifferenceHandler.getTextFromTranslation("config." + ETFClientCommon.MOD_ID + ".reset_defaults"),
                 (button) -> {