+++ common/src/main/java/traben/entity_texture_features/mixin/entity/feature_renderers/MixinElytraFeatureRenderer.java
@@ -1,4 +1,4 @@
-package traben.entity_texture_features.mixin.entity.featureRenderers;
+package traben.entity_texture_features.mixin.entity.feature_renderers;
 
 import com.google.common.collect.ImmutableList;
 import net.minecraft.client.MinecraftClient;