+++ forge/src/main/java/traben/entity_texture_features/forge/ETFVersionDifferenceHandlerImpl.java
@@ -10,6 +10,7 @@ import org.slf4j.LoggerFactory;
 
 import java.io.File;
 
+@SuppressWarnings("SameReturnValue")
 public class ETFVersionDifferenceHandlerImpl {